+++ src/components/organisms/index.ts
@@ -1,5 +1,5 @@
 import InputTest from './inputTest/InputTest';
 import Header from './header/Header';
-import AddTaskDialog from './addTeskDialog/AddTaskDialog';
+import AddTaskDialog from './teskDialog/TaskDialog';
 
 export { InputTest, Header, AddTaskDialog };
