+++ src/components/organisms/teskDialog/data/taskDialog.types.ts
@@ -1,0 +1,6 @@
+import { Task } from '@/types/task/task.type';
+
+export type TaskDialogProps = {
+  dictionary?: {};
+  task?: Task;
+};
