+++ src/libs/redux/slices/index.ts
@@ -1,1 +1,2 @@
 export * from './counterSlice';
+export * from './taskSlice';
